% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_delete_resource_definition}
\alias{greengrass_delete_resource_definition}
\title{Deletes a resource definition}
\usage{
greengrass_delete_resource_definition(ResourceDefinitionId)
}
\arguments{
\item{ResourceDefinitionId}{[required] The ID of the resource definition.}
}
\description{
Deletes a resource definition.
}
\section{Request syntax}{
\preformatted{greengrass$delete_resource_definition(
  ResourceDefinitionId = "string"
)
}
}

\keyword{internal}
