% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_associated_role}
\alias{greengrass_get_associated_role}
\title{Retrieves the role associated with a particular group}
\usage{
greengrass_get_associated_role(GroupId)
}
\arguments{
\item{GroupId}{[required] The ID of the Greengrass group.}
}
\description{
Retrieves the role associated with a particular group.
}
\section{Request syntax}{
\preformatted{greengrass$get_associated_role(
  GroupId = "string"
)
}
}

\keyword{internal}
