% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_deployment_status}
\alias{greengrass_get_deployment_status}
\title{Returns the status of a deployment}
\usage{
greengrass_get_deployment_status(DeploymentId, GroupId)
}
\arguments{
\item{DeploymentId}{[required] The ID of the deployment.}

\item{GroupId}{[required] The ID of the Greengrass group.}
}
\description{
Returns the status of a deployment.
}
\section{Request syntax}{
\preformatted{greengrass$get_deployment_status(
  DeploymentId = "string",
  GroupId = "string"
)
}
}

\keyword{internal}
