% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_resource_definition}
\alias{greengrass_get_resource_definition}
\title{Retrieves information about a resource definition, including its
creation time and latest version}
\usage{
greengrass_get_resource_definition(ResourceDefinitionId)
}
\arguments{
\item{ResourceDefinitionId}{[required] The ID of the resource definition.}
}
\description{
Retrieves information about a resource definition, including its creation time and latest version.
}
\section{Request syntax}{
\preformatted{greengrass$get_resource_definition(
  ResourceDefinitionId = "string"
)
}
}

\keyword{internal}
