% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_service_role_for_account}
\alias{greengrass_get_service_role_for_account}
\title{Retrieves the service role that is attached to your account}
\usage{
greengrass_get_service_role_for_account()
}
\description{
Retrieves the service role that is attached to your account.
}
\section{Request syntax}{
\preformatted{greengrass$get_service_role_for_account()
}
}

\keyword{internal}
