% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_list_group_versions}
\alias{greengrass_list_group_versions}
\title{Lists the versions of a group}
\usage{
greengrass_list_group_versions(GroupId, MaxResults, NextToken)
}
\arguments{
\item{GroupId}{[required] The ID of the Greengrass group.}

\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Lists the versions of a group.
}
\section{Request syntax}{
\preformatted{greengrass$list_group_versions(
  GroupId = "string",
  MaxResults = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
