% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_update_connectivity_info}
\alias{greengrass_update_connectivity_info}
\title{Updates the connectivity information for the core}
\usage{
greengrass_update_connectivity_info(ConnectivityInfo, ThingName)
}
\arguments{
\item{ConnectivityInfo}{A list of connectivity info.}

\item{ThingName}{[required] The thing name.}
}
\description{
Updates the connectivity information for the core. Any devices that belong to the group which has this core will receive this information in order to find the location of the core and connect to it.
}
\section{Request syntax}{
\preformatted{greengrass$update_connectivity_info(
  ConnectivityInfo = list(
    list(
      HostAddress = "string",
      Id = "string",
      Metadata = "string",
      PortNumber = 123
    )
  ),
  ThingName = "string"
)
}
}

\keyword{internal}
