% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_update_resource_definition}
\alias{greengrass_update_resource_definition}
\title{Updates a resource definition}
\usage{
greengrass_update_resource_definition(Name, ResourceDefinitionId)
}
\arguments{
\item{Name}{The name of the definition.}

\item{ResourceDefinitionId}{[required] The ID of the resource definition.}
}
\description{
Updates a resource definition.
}
\section{Request syntax}{
\preformatted{greengrass$update_resource_definition(
  Name = "string",
  ResourceDefinitionId = "string"
)
}
}

\keyword{internal}
